telegram-bot/requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
aiosqlite==0.19.0
openai==1.3.5